Added routes for get and delete postbyid

+++ data/post.js
@@ -1,32 +1,45 @@
 const db = require("../config");
 
 const posts = db.postsCollection;
-const errorHandling = require("../validations/dataValidations");
-const validations = errorHandling.userValidations;
+const validations = require("../validations/dataValidations");
+//const validations = errorHandling.userValidations;
 const { ObjectId } = require("mongodb");
 
 /**Database function for the Post Collection */
-const getPostById = async (PostId) => {
-  // if (!PostId) {
-  //   throw 'You must provide an id to search for';
-  // }
-  // if (typeof PostId !== 'string') {
-  //   throw 'Id must be a string';
-  // }
-  // if (PostId.trim().length === 0) {
-  //   throw 'Id cannot be an empty string or just spaces';
-  // }
-  // id = PostId.trim();
-  // if (!ObjectId.isValid(id)) {
-  //   throw 'invalid object ID';
-  // }
-  // const postCollection = await posts();
-  // const post = await postCollection.findOne({ _id: ObjectId(id)});
-  // if (post === null) {
-  //   throw 'no post with that id';
-  // }
-  // post._id = post._id.toString();
-  // return post;
+//const getPostById = async (PostId) => {
+// if (!PostId) {
+//   throw 'You must provide an id to search for';
+// }
+// if (typeof PostId !== 'string') {
+//   throw 'Id must be a string';
+// }
+// if (PostId.trim().length === 0) {
+//   throw 'Id cannot be an empty string or just spaces';
+// }
+// id = PostId.trim();
+// if (!ObjectId.isValid(id)) {
+//   throw 'invalid object ID';
+// }
+// const postCollection = await posts();
+// const post = await postCollection.findOne({ _id: ObjectId(id)});
+// if (post === null) {
+//   throw 'no post with that id';
+// }
+// post._id = post._id.toString();
+// return post;
+//};
+const getPostById = async (postId) => {
+  //validation of id
+  validations.validateID(postId)
+  postId = postId.trim()
+  const postCollection = await posts();
+  const postInfo = await postCollection.findOne({ _id: ObjectId(postId) });
+  if (postInfo === null) {
+    throw 'There is no post for the given id';
+  }
+  postInfo._id = postInfo._id.toString();
+  console.log(postInfo)
+  return postInfo;
 };
 
 const addPost = async (
@@ -89,28 +102,22 @@ const getAllPosts = async () => {
   return PostList;
 };
 
-const removePost = async (PostId) => {
-  // if (!PostId) {
-  //   throw 'You must provide an id to search for';
-  // }
-  // if (typeof PostId !== 'string') {
-  //   throw 'Id must be a string';
-  // }
-  // if (PostId.trim().length === 0) {
-  //   throw 'id cannot be an empty string or just spaces';
-  // }
-  // id = PostId.trim();
-  // if (!ObjectId.isValid(id)) {
-  //   throw 'invalid object ID';
-  // }
+const removePost = async (postId) => {
+  //validation of id
+  validations.validateID(postId)
+  id = postId.trim()
   const postCollection = await posts();
+  const postInfo = await postCollection.findOne({ _id: ObjectId(id) });
+  if (!postInfo) {
+    throw "There is no post for the given id";
+  }
   const deletionInfo = await postCollection.deleteOne({ _id: ObjectId(id) });
-
   if (deletionInfo.deletedCount === 0) {
     throw `Could not delete post with id of ${id}`;
   }
   return { deleted: true };
 };
+
 module.exports = {
   getPostById,
   addPost,
